Move out of firehose folder
