+++ src/stackdriver-nozzle/filter/filter_test.go
@@ -2,7 +2,7 @@ package filter_test
 
 import (
 	"github.com/cloudfoundry/sonde-go/events"
-	"github.com/evandbrown/gcp-tools-release/src/stackdriver-nozzle/firehose/filter"
+	"github.com/evandbrown/gcp-tools-release/src/stackdriver-nozzle/filter"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
